requirements.txt
flask
pytest
pytest-cov
pytest-mock
selenium
behave
phantomjs-binary